% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sneat_button.R
\name{sneat_button}
\alias{sneat_button}
\title{sneat template for action button}
\usage{
sneat_button(inputId, label, icon = NULL, width = NULL, disabled = FALSE, ...)
}
\description{
sneat template for action button
}
