% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sneat_select.R
\name{sneat_select}
\alias{sneat_select}
\title{sneat template for select input}
\usage{
sneat_select(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE,
  width = NULL,
  size = NULL
)
}
\description{
sneat template for select input
}
